Move auth to dialler pkg
