+++ pkg/dialler/auth.go
@@ -1,4 +1,4 @@
-package auth
+package dialler
 
 import (
 	"context"
